+++ frontend/src/requests/Requester.js
@@ -1,3 +1,4 @@
 // move the variable value to env when the web it's on production
 export const SERVER_DOMAIN = 'http://localhost:4000'
 
+export const AUTH_DOMAIN = "http://localhost:4500";
